wireguard: fix up faulty snapshot patcher
I forgot to update the paths in create-patch.sh of the latest snapshot,
which means this will likely produce a bad patch. So, fixing it up here
as a stop-gap measure, so that Sultan doesn't freak out when my patch
ruins his kernel compiles.

This should be reverted as soon as a new snapshot upstream is released.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>